rename(manageProject): rename components
