+++ src/frontend/src/components/ManageProject/Details/validation/EditDetailsValidation.ts
